requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
idna==3.10
mysql-connector-python==9.4.0
psutil==7.0.0
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
